libbio: add casts to eliminate -Wconversion warning
LGTM=bradfitz
R=golang-codereviews, bradfitz
CC=golang-codereviews
https://golang.org/cl/87140044

+++ src/libbio/bgetc.c
@@ -83,7 +83,7 @@ Bgetle4(Biobuf *bp)
 
 	l = Bgetle2(bp);
 	h = Bgetle2(bp);
-	return l|((uint32)h<<16);
+	return (int)((uint32)l|((uint32)h<<16));
 }
 
 int
